+++ tests/integration/client/test_other_user.py
@@ -1,8 +1,9 @@
-from deprecated import deprecated
 import pytest
+from deprecated import deprecated
 
 from specklepy.api.client import SpeckleClient
 from specklepy.api.models import Activity, ActivityCollection, LimitedUser
+
 
 @deprecated()
 @pytest.mark.run(order=4)
